256_color_v2/requirements.txt
imageio
python-chess
CairoSVG==2.1.3
